+++ src/test/java/ch/bitagent/bitcoin/lib/wallet/WalletTest.java
@@ -79,8 +79,7 @@ class WalletTest {
 
     @Disabled(value = "manual")
     @Test
-    void
-    createSegwitTx() {
+    void createNativeSegwitTx() {
         var mnemonicSentence = Properties.getWalletMnemonic(WALLET_DEV_FILENAME, 0);
         var wallet = Wallet.parse(mnemonicSentence, null, Wallet.PURPOSE_NATIVE_SEGWIT, Wallet.COIN_TYPE_BITCOIN, 0, 3);
         wallet.history(0);
@@ -88,19 +87,13 @@ class WalletTest {
         log.info(wallet.toString());
         assertFalse(wallet.getUtxoList().isEmpty());
 
-        long utxoAmount = 0L;
-        List<TxIn> utxoList = new ArrayList<>();
-        for (Utxo utxo : wallet.getUtxoList()) {
-            log.info(utxo.toString());
-            if (utxo.getHeight() > 0) {
-                utxoList.add(new TxIn(utxo));
-                utxoAmount += utxo.getValue();
-            }
-        }
+        List<TxIn> txInList = new ArrayList<>();
+        long utxoAmount = getUtxoAmount(wallet, txInList);
         assertTrue(utxoAmount > 0);
 
+        var spendAmount = utxoAmount;
         var spendAddress = wallet.nextReceiveAddress();
-        var spendTxOut = new TxOut(Int.parse(utxoAmount), spendAddress.scriptPubkey());
+        var spendTxOut = new TxOut(Int.parse(spendAmount), spendAddress.scriptPubkey());
         var opReturnTxOut = new TxOut(Int.parse(0), Script.opReturn("bitcoinjavalib"));
 
         var version = 2;
@@ -108,27 +101,92 @@ class WalletTest {
         var heigth = electrum.height();
         Map<String, String> cache = new HashMap<>();
 
-        var tx = new Tx(Int.parse(version), utxoList, List.of(spendTxOut, opReturnTxOut), Int.parse(heigth), false, true);
-        txSignInput(wallet, tx, utxoList, cache);
-        log.info(String.format("size %sB/%swu/%svB", tx.sizeBytes(), tx.sizeWeightUnits(), tx.sizeVirtualBytes()));
+        var tx = new Tx(Int.parse(version), txInList, List.of(spendTxOut, opReturnTxOut), Int.parse(heigth), false, true);
+        txSignInput(wallet, tx, txInList, cache);
 
         var satsVB = electrum.estimateFee(1);
         var feeAmount = tx.sizeVirtualBytes() * satsVB;
-        spendTxOut = new TxOut(Int.parse(utxoAmount - feeAmount), spendAddress.scriptPubkey());
-        tx = new Tx(Int.parse(version), utxoList, List.of(spendTxOut, opReturnTxOut), Int.parse(heigth), false, true);
-        txSignInput(wallet, tx, utxoList, cache);
-        log.info(String.format("fee %ssatsVB/%ssats", satsVB, feeAmount));
+        spendAmount = utxoAmount - feeAmount;
+        spendTxOut = new TxOut(Int.parse(spendAmount), spendAddress.scriptPubkey());
+        tx = new Tx(Int.parse(version), txInList, List.of(spendTxOut, opReturnTxOut), Int.parse(heigth), false, true);
+        txSignInput(wallet, tx, txInList, cache);
         assertTrue(feeAmount > 0);
         assertEquals(feeAmount, tx.fee(cache).intValue());
         log.info(tx.toString());
 
+        log.info(String.format("tx %svB", tx.sizeVirtualBytes()));
+        log.info(String.format("%s sats/vB", satsVB));
+        log.info(String.format("utxo %s", utxoAmount));
+        log.info(String.format("spend %s", spendAmount));
+        log.info(String.format("fee %s", feeAmount));
         log.info(String.format("broadcast transaction%n%s", tx.hexString()));
     }
 
-    private void txSignInput(Wallet wallet, Tx tx, List<TxIn> utxoList, Map<String, String> cache) {
-        for (int i = 0; i < utxoList.size(); i++) {
-            assertTrue(tx.signInput(i, wallet.getPrivateKeyForChangeIndex(utxoList.get(i).getUtxo().getChangeIndex()), cache));
+    private long getUtxoAmount(Wallet wallet, List<TxIn> txInList) {
+        long utxoAmount = 0;
+        for (Utxo utxo : wallet.getUtxoList()) {
+            log.info(utxo.toString());
+            if (utxo.getHeight() > 0) {
+                txInList.add(new TxIn(utxo));
+                utxoAmount += utxo.getValue();
+            }
         }
+        return utxoAmount;
+    }
+
+    private void txSignInput(Wallet wallet, Tx tx, List<TxIn> txInList, Map<String, String> cache) {
+        for (int i = 0; i < txInList.size(); i++) {
+            assertTrue(tx.signInput(i, wallet.getPrivateKeyForChangeIndex(txInList.get(i).getUtxo().getChangeIndex()), cache));
+        }
+    }
+
+    @Disabled(value = "manual")
+    @Test
+    void createChangeTx() {
+        var mnemonicSentence = Properties.getWalletMnemonic(WALLET_DEV_FILENAME, 0);
+        var wallet = Wallet.parse(mnemonicSentence, null, Wallet.PURPOSE_NATIVE_SEGWIT, Wallet.COIN_TYPE_BITCOIN, 0, 3);
+        wallet.history(0);
+        wallet.history(1);
+        log.info(wallet.toString());
+        assertFalse(wallet.getUtxoList().isEmpty());
+
+        List<TxIn> txInList = new ArrayList<>();
+        long utxoAmount = getUtxoAmount(wallet, txInList);
+        assertTrue(utxoAmount > 0);
+
+        // tested with p2pkh address 1JQwARc3U8GKwsiuV42VPCAATsCig7Lh15
+        // tested with p2sh address 3KeHaLZ6GFMYoZ37HpzixVqsn2FFxkNi3D
+        // not tested yet with p2tr address bc1p36065kzufa0d6wm9ay59fz868rwaqauhffnhm0m4a3sdd4lqswlqs3dut6
+        var spendAmount = 10000;
+        var spendAddress = Address.parse("bc1p36065kzufa0d6wm9ay59fz868rwaqauhffnhm0m4a3sdd4lqswlqs3dut6");
+        var spendTxOut = new TxOut(Int.parse(spendAmount), spendAddress.scriptPubkey());
+        var changeAmount = utxoAmount - spendAmount;
+        var changeAddress = wallet.nextChangeAddress();
+        var changeTxOut = new TxOut(Int.parse(changeAmount), changeAddress.scriptPubkey());
+
+        var version = 2;
+        var electrum = new Electrum();
+        var heigth = electrum.height();
+        Map<String, String> cache = new HashMap<>();
+
+        var tx = new Tx(Int.parse(version), txInList, List.of(spendTxOut, changeTxOut), Int.parse(heigth), false, true);
+        txSignInput(wallet, tx, txInList, cache);
+        var satsVB = electrum.estimateFee(1);
+
+        var feeAmount = tx.sizeVirtualBytes() * satsVB;
+        changeAmount = utxoAmount - spendAmount - feeAmount;
+        changeTxOut = new TxOut(Int.parse(changeAmount), changeAddress.scriptPubkey());
+        tx = new Tx(Int.parse(version), txInList, List.of(spendTxOut, changeTxOut), Int.parse(heigth), false, true);
+        txSignInput(wallet, tx, txInList, cache);
+        log.info(tx.toString());
+
+        log.info(String.format("tx %svB", tx.sizeVirtualBytes()));
+        log.info(String.format("%s sats/vB", satsVB));
+        log.info(String.format("utxo %s", utxoAmount));
+        log.info(String.format("spend %s", spendAmount));
+        log.info(String.format("fee %s", feeAmount));
+        log.info(String.format("change %s", changeAmount));
+        log.info(String.format("broadcast transaction%n%s", tx.hexString()));
     }
 
     @Disabled(value = "manual")